requirements.txt
discord.py==0.16.12
jsonschema==2.6.0